[ghc8] Allow XML tags brought in by external namespaces
Many of our idl interfaces utilize the DBus spec, which introduces
tags that live in the "tp:" namespace. The upgraded dbus library
does not support parsing these tags. This patch ensures we don't
fail early if we come across a tag we can't process at the node
level. Ignoring tags in nested levels, such as in methods, prevents
the method args from being parsed and therefore rpc-proxy fails
to build up the correct DBus call when introspecting the bus. This
causes methods to be called without arguments, resulting in many
failures.

Note that this does imply that the tp: tags are ignored, but this
change seemingly has no bearing on the generated stubs; the enums
are present despite the new behavior. The rpcgen code that parses
the enum XML tags and generates the stubs is separate from the
DBus xml parse, and it is tailor made to understand the enum
tags.

Signed-off-by: Nicholas Tsirakis <tsirakisn@ainfosec.com>
Signed-off-by: Chris Rogers <rogersc@ainfosec.com>